presence_prediction/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
convertdate==2.4.0
cycler==0.11.0
fonttools==4.33.3
geographiclib==1.52
geopy==2.2.0
googlemaps==4.6.0
hijri-converter==2.2.4
holidays==0.14.2
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
matplotlib==3.5.2
numpy==1.22.4
packaging==24.1
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.1
PyMeeus==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scipy==1.8.1
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
typing_extensions==4.2.0
urllib3==1.26.9
torch==2.1.2
pytorch-lightning==2.1.3
scikit-learn==1.1.3
google-cloud-storage==2.16.0